+++ src/lib/common/XmppQueue/XmppClient.ts
@@ -2,6 +2,7 @@
 //import Element from "ltx";
 import {NameSpacesLabels} from "../../connection/XMPPService";
 import {DataStoreType} from "../../config/config";
+import {retry} from "rxjs";
 
 export {};
 
@@ -27,6 +28,7 @@ const _streamFeatures = require('@xmpp/stream-features');
 const plain = require('@xmpp/sasl-plain');
 const xml = require("@xmpp/xml");
 //const debug = require("@xmpp/debug");
+const xml2js = require('xml2js');
 
 const Element = require('ltx').Element;
 
@@ -47,6 +49,7 @@ class XmppClient  {
 	public password: any;
     socketClosed: boolean = false;
     storeMessages: any;
+    enablesendurgentpushmessages: any;
     copyMessage: any = true;
     rateLimitPerHour: any;
     private nbMessagesSentThisHour: number;
@@ -79,7 +82,7 @@ class XmppClient  {
 
     }
 
-    init(_logger, _eventemitter, _timeBetweenXmppRequests, _storeMessages, _rateLimitPerHour, _messagesDataStore, _copyMessage) {
+    init(_logger, _eventemitter, _timeBetweenXmppRequests, _storeMessages, _rateLimitPerHour, _messagesDataStore, _copyMessage, _enablesendurgentpushmessages) {
         let that = this;
         that.logger = _logger;
         this.eventEmitter = _eventemitter;
@@ -90,6 +93,7 @@ class XmppClient  {
         that.messagesDataStore = _messagesDataStore;
         that.lastTimeReset = new Date ();
         that.copyMessage = _copyMessage;
+        that.enablesendurgentpushmessages = _enablesendurgentpushmessages;
 
         if (that.messagesDataStore) {
             switch (that.messagesDataStore) {
@@ -200,12 +204,25 @@ class XmppClient  {
         that.logger.log("debug", LOG_ID + "(resetnbMessagesSentThisHour) _exiting_");
     }
 
+    async getJsonFromXML(xml : string) {
+        try {
+            const result = await xml2js.parseStringPromise(xml, {mergeAttrs: false, explicitArray : false, attrkey : "$attrs", emptyTag  : undefined});
+
+            // convert it to a JSON string
+            return result;
+            //return JSON.stringify(result, null, 4);
+        } catch (err) {
+            //console.log(err);
+            return {};
+        }
+    }
+    
     send(...args) {
         let that = this;
         that.logger.log("debug", LOG_ID + "(send) _entering_");
         return new Promise((resolve) => {
             let prom = this.xmppQueue.addPromise(
-                new Promise((resolve2, reject2) => {
+                new Promise(async (resolve2, reject2) => {
                     /*
                     if (args && args[0]) {
                         that.logger.log("internal", LOG_ID + "(send) stanza to send ", that.logger.colors.gray(args[0].toString()));
@@ -222,6 +239,23 @@ class XmppClient  {
 
                     let stanza = args[0];
 
+                    if (that.enablesendurgentpushmessages && stanza && stanza.name == "message") {
+                        let stanzaJson = await that.getJsonFromXML(stanza);
+                        that.logger.log("internal", LOG_ID + "(send) enablesendurgentpushmessages is setted, and of type message, JSONstanza is : ", stanzaJson);
+                        //if (stanzaJson && stanzaJson.message != undefined) {
+                            //that.logger.log("internal", LOG_ID + "(send) enablesendurgentpushmessages is setted, stanza of type message.");
+                            if (stanzaJson.message.body && stanzaJson.message.body != "") {
+                                that.logger.log("internal", LOG_ID + "(send) enablesendurgentpushmessages is setted, stanza of type message with not empty body.");
+                                // <retry-push xmlns='urn:xmpp:hints'/> 
+                                let retryPush = "retry-push"; 
+                                stanza.append(xml(retryPush, {
+                                    "xmlns": NameSpacesLabels.HintsNameSpace
+                                }));
+                                that.logger.log("internal", LOG_ID + "(send) enablesendurgentpushmessages is setted, stanza of type message with not empty body.");
+                            } 
+                        //} 
+                    }
+                    
                     if (that.storeMessages == false && stanza && typeof stanza === "object" && stanza.name == "message") {
                    // if (that.storeMessages == false && stanza && typeof stanza === "object" && stanza.name == "message") {
                         // that.logger.log("info", LOG_ID + "(send) will add <no-store /> to stanza.");